# api/openapi.yaml
openapi: "3.1.0"
info:
  title: EncodeSwarmr API
  version: "1.0.0"
  description: >
    REST API for the EncodeSwarmr controller. Manages encoding jobs,
    agents, sources, templates, variables, webhooks, users, and audit logs.
    All versioned endpoints are prefixed with /api/v1.
servers:
  - url: /api/v1
    description: Versioned API base path
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      summary: List all agents
      tags: [agents]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Agent"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /agent-tokens:
    get:
      operationId: listEnrollmentTokens
      summary: List enrollment tokens
      description: Returns all tokens (without the token value). Requires admin role.
      tags: [agent-tokens]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Enrollment token list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/EnrollmentToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createEnrollmentToken
      summary: Create an enrollment token
      description: >
        Creates a one-time token for enrolling a new agent. The token value is
        only returned in this response and cannot be retrieved again.
        Requires admin role.
      tags: [agent-tokens]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expires_in:
                  type: string
                  description: Go duration string (e.g. 24h, 72h). Defaults to 24h.
                  example: "48h"
      responses:
        "201":
          description: Token created (token value shown once)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/EnrollmentTokenCreate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /agent-tokens/{id}:
    delete:
      operationId: deleteEnrollmentToken
      summary: Delete an enrollment token
      description: Requires admin role.
      tags: [agent-tokens]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Token deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Sources ───────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      summary: List users
      description: Requires admin role.
      tags: [users]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createUser
      summary: Create a user
      description: Requires admin role.
      tags: [users]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: HTTP-only session cookie set by POST /auth/login or OIDC callback.
    bearerAuth:
      type: http
      scheme: bearer
      description: API key sent as a Bearer token in the Authorization header.

  # ─── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Insufficient role
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Validation error (unprocessable entity)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    Envelope:
      type: object
      properties:
        data:
          description: The primary response payload.
        meta:
          $ref: "#/components/schemas/Meta"
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
          description: Correlation identifier echoed from X-Request-ID.
...
    Problem:
      description: RFC 9457 Problem Details
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string

    # Domain schemas
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        hostname:
          type: string
        ip_address:
          type: string
        status:
          type: string
          enum: [idle, running, offline, draining, pending_approval]
        tags:
          type: array
          items:
            type: string
          nullable: true
        version:
          type: string
          nullable: true
        vnc_port:
          type: integer
          nullable: true
        current_task_id:
          type: string
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [viewer, operator, admin]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EnrollmentToken:
      type: object
      description: Enrollment token record (token value omitted).
      properties:
        id:
          type: string
        used_by:
          type: string
          nullable: true
          description: Agent ID that consumed this token.
        used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    EnrollmentTokenCreate:
      allOf:
        - $ref: "#/components/schemas/EnrollmentToken"
        - type: object
          properties:
            token:
              type: string
              description: The one-time token value. Only returned at creation time.
...
    CreateUserRequest:
      type: object
      required: [username, email, role]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [viewer, operator, admin]
        password:
          type: string
          format: password
          description: Optional for OIDC-only users.